+++ gems/aws-sdk-core/CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased Changes
 ------------------
 
+* Issue - Ensure that nested sensitive members and sensitive members of lists and maps are filtered.
+
 3.201.1 (2024-07-05)
 ------------------